chatbot-backend/requirements.txt
fastapi==0.103.1
Markdown==3.4.4
networkx==3.1
pydantic==2.3.0
pydantic_settings==2.0.3
PyJWT==2.8.0
pytest==7.4.2
python-dotenv==1.0.0
uvicorn==0.23.2
sentencepiece==0.1.99
protobuf==4.24.4
websockets==11.0.3
